pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-sre-doctor"
version = "0.1.0"
description = "CLI profissional para troubleshooting operacional em AWS, com foco em diagnóstico rápido e redução de MTTR."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Tharlesson" }]
dependencies = [
  "boto3>=1.35.0,<2.0.0",
  "typer>=0.12.0,<1.0.0",
  "rich>=13.9.0,<14.0.0",
  "structlog>=24.0.0,<25.0.0",
  "PyYAML>=6.0.0,<7.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0"
]

[project.scripts]
aws-sre-doctor = "cli.main:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["cli*", "core*", "checks*", "reporters*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
